# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvestPro Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Users
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/transactions:
    get:
      operationId: listTransactions
      tags: [users]
      summary: List all transactions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"

  # Dashboard
...
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List user investments
      responses:
        "200":
          description: Investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create a new investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Investment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          description: Validation error
  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get investment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"

  # Deposits
  /deposits:
    get:
      operationId: listDeposits
      tags: [deposits]
      summary: List user deposits
      responses:
        "200":
          description: Deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Request a deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/{id}:
    get:
      operationId: getDeposit
      tags: [deposits]
      summary: Get deposit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"

  # Withdrawals
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [firstName, lastName, email, phone, password, country]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        country:
          type: string
        city:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # User
    User:
      type: object
      required: [id, firstName, lastName, email, phone, country, role, status, kycLevel, vipLevel, referralCode, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, banned]
        kycLevel:
          type: integer
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        vipLevel:
          type: integer
        referralCode:
          type: string
        totalInvested:
          type: number
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        city:
          type: string
        walletAddress:
          type: ["string", "null"]

    # Wallet
...
    Transaction:
      type: object
      required: [id, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, earning, bonus, commission]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, rejected]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Dashboard
...
    Investment:
      type: object
      required: [id, planId, planName, amount, dailyRate, durationDays, startDate, endDate, status, totalEarned]
      properties:
        id:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        amount:
          type: number
        dailyRate:
          type: number
        durationDays:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
        totalEarned:
          type: number
        dailyEarning:
          type: number
    InvestmentInput:
      type: object
      required: [planId, amount]
      properties:
        planId:
          type: integer
        amount:
          type: number

    # Deposits
    Deposit:
      type: object
      required: [id, amount, status, walletAddress, txHash, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        walletAddress:
          type: string
        txHash:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
    DepositInput:
      type: object
      required: [amount, txHash]
      properties:
        amount:
          type: number
        txHash:
          type: string
        walletAddress:
          type: string

    # Withdrawals
...